requirements.txt
asgiref==3.6.0
avinit==1.3.0
bcrypt==4.0.1
boto3==1.34.64
botocore==1.34.64
CacheControl==0.14.0
cachetools==5.3.3
cairocffi==1.6.1
CairoSVG==2.7.1
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==3.1.0
cryptography==42.0.8
cssselect2==0.7.0
defusedxml==0.7.1
Django==4.1.7
django-cors-headers==3.14.0
django-filter==22.1
django-material-admin==1.8.6
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
firebase-admin==6.5.0
google-api-python-client==2.132.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
google-api-core
google-cloud-firestore
googleapis-common-protos
grpcio-status
mysql-connector-python
protobuf
grpcio==1.64.1
httplib2==0.22.0
idna==3.4
jmespath==1.0.1
msgpack==1.0.8
mysqlclient==2.1.1
Pillow==10.0.0
proto-plus==1.23.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
pyfcm==1.5.4
PyJWT==2.6.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2022.7.1
requests==2.28.2
rest-framework-simplejwt==0.0.2
rsa==4.9
s3transfer==0.10.1
six==1.16.0
sqlparse==0.4.3
tinycss2==1.2.1
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
webencodings==0.5.1
